# ASV: benchmark for DataFrame.Update