--- a/asv_bench/benchmarks/frame_methods.py
+++ b/asv_bench/benchmarks/frame_methods.py
@@ -865,14 +865,16 @@ class FindValidIndex:
 class Update:
     def setup(self):
         rng = np.random.default_rng()
-        self.df = DataFrame(rng.uniform(size=(1000, 10)))
+        self.df = DataFrame(rng.uniform(size=(10_000, 10)))
 
-        idx = rng.choice(range(1000), size=1000, replace=False)
+        idx = rng.choice(range(10_000), size=10_000, replace=False)
         self.df_random = DataFrame(self.df, index=idx)
 
-        idx = rng.choice(range(1000), size=100, replace=False)
+        idx = rng.choice(range(10_000), size=1_000, replace=False)
         cols = rng.choice(range(10), size=2, replace=False)
-        self.df_sample = DataFrame(rng.uniform(size=(100, 2)), index=idx, columns=cols)
+        self.df_sample = DataFrame(
+            rng.uniform(size=(1_000, 2)), index=idx, columns=cols
+        )
 
     def time_to_update_big_frame_small_arg(self):
         self.df.update(self.df_sample)