- [ ] closes #xxxx (Replace xxxx with the GitHub issue number)
- [ ] [Tests added and passed](https://pandas.pydata.org/pandas-docs/dev/development/contributing_codebase.html#writing-tests) if fixing a bug or adding a new feature
- [ ] All [code checks passed](https://pandas.pydata.org/pandas-docs/dev/development/contributing_codebase.html#pre-commit).
- [ ] Added [type annotations](https://pandas.pydata.org/pandas-docs/dev/development/contributing_codebase.html#type-hints) to new arguments/methods/functions.
- [ ] Added an entry in the latest `doc/source/whatsnew/vX.X.X.rst` file if fixing a bug or adding a new feature.

@rhshadrach and @datapythonista 

Benchmarks for the DataFrame.Update method are implemented, as suggested on PR #57637 by @datapythonista and based on the manual tests done by @rhshadrach on #55634. By running ASV locally we measure the change in the performance due to the fix on #57637, as indicated below:

| Change   | Main (2.2.1) [8da8b544]   | PR (3.0) [93ea131f]  |   Ratio | Benchmark (Parameter)                                   |
|----------|----------------------------|----------------------------------------|---------|---------------------------------------------------------|
| -        | 1.13±0.02ms                | 887±10μs                               |    0.78 | frame_methods.Update.time_to_update_random_indices      |
| -        | 1.19±0.02ms                | 840±9μs                                |    0.71 | frame_methods.Update.time_to_update_small_frame_big_arg |
| -        | 1.27±0.01ms                | 859±9μs                                |    0.67 | frame_methods.Update.time_to_update_big_frame_small_arg |

Regards