@aureliobarbosa 2024-04-13
> Can you increase the parameters to get runtime up to ~10ms.

@rhshadrach Think I am done here.

Data was increased twofold, from 10**4 to 10**6 points. Each test is now taking ~4ms. Didn't check the scaling, but it seems to be enough because the data size is similar to other benchmarks in the `frame_methods.py`. If needed, I can increase the data again. 

It would be interesting to have some guidance on the best data size for a useful test, but that would be a whole work on itself.